clean up remaining constants
